+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.0.2] - 2022-08-02
 ### Changed
 - Change `name` in pyproject.toml from `origami` to `noteable-origami`
+- Disable tag generation by bumpversion
 
 ### Added
 - Add `packages` to pyproject.toml