+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fix `SessionRequestDetails.generate_file_request` handling of file paths with folders
+
 ## [0.0.2] - 2022-08-02
 ### Changed
 - Change `name` in pyproject.toml from `origami` to `noteable-origami`