+++ CHANGELOG.md
@@ -5,12 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.0.3] - 2022-09-02
 ### Fixed
 - Fix `SessionRequestDetails.generate_file_request` handling of file paths with folders
 
 ### Added
 - Add ability to delete/add cells with NoteableClient
-
 
 ## [0.0.2] - 2022-08-02
 ### Changed